pull classes out into separate modules
